Clang tidy

+++ source/IOHaptics/include/IOJson.h
@@ -53,28 +53,28 @@ namespace haptics::io {
 
 constexpr int BYTE_SIZE_IO = 8;
 
-constexpr short COMPARE_START = 32;
-constexpr short ASCII_UPPER_1 = 65;
-constexpr short ASCII_UPPER_2 = 90;
-constexpr short ASCII_LOWER_1 = 97;
-constexpr short ASCII_LOWER_2 = 122;
-constexpr short ASCII_DIGIT_1 = 48;
-constexpr short ASCII_DIGIT_2 = 57;
-constexpr short ASCII_PLUS = 43;
-constexpr short ASCII_SOLIDUS = 47;
+constexpr int COMPARE_START = 32;
+constexpr int ASCII_UPPER_1 = 65;
+constexpr int ASCII_UPPER_2 = 90;
+constexpr int ASCII_LOWER_1 = 97;
+constexpr int ASCII_LOWER_2 = 122;
+constexpr int ASCII_DIGIT_1 = 48;
+constexpr int ASCII_DIGIT_2 = 57;
+constexpr int ASCII_PLUS = 43;
+constexpr int ASCII_SOLIDUS = 47;
 
-constexpr short BASE64_UPPER_1 = 0;
-constexpr short BASE64_UPPER_2 = 25;
-constexpr short BASE64_LOWER_1 = 26;
-constexpr short BASE64_LOWER_2 = 51;
-constexpr short BASE64_DIGIT_1 = 52;
-constexpr short BASE64_DIGIT_2 = 61;
-constexpr short BASE64_PLUS = 62;
-constexpr short BASE64_SOLIDUS = 63;
+constexpr int BASE64_UPPER_1 = 0;
+constexpr int BASE64_UPPER_2 = 25;
+constexpr int BASE64_LOWER_1 = 26;
+constexpr int BASE64_LOWER_2 = 51;
+constexpr int BASE64_DIGIT_1 = 52;
+constexpr int BASE64_DIGIT_2 = 61;
+constexpr int BASE64_PLUS = 62;
+constexpr int BASE64_SOLIDUS = 63;
 
-constexpr short DIFF_UPPER = 65;
-constexpr short DIFF_LOWER = 71;
-constexpr short DIFF_DIGIT = 4;
+constexpr int DIFF_UPPER = 65;
+constexpr int DIFF_LOWER = 71;
+constexpr int DIFF_DIGIT = 4;
 
 constexpr size_t BASE64_SIZE = 6;
